Reorganize tests out of EventCreatingTests and EventDecodingTests into event kind specific tests
