+++ Tests/NostrSDKTests/Events/Calendars/CalendarEventParticipantTests.swift
